backend/requirements.txt
# ── Core Framework ─────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-dotenv==1.0.1

# ── Database ──────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# ── File Handling ─────────────────────────────────────
python-multipart==0.0.20
aiofiles==24.1.0

# ── Vector Store ──────────────────────────────────────
qdrant-client==1.12.1

# ── LLM ───────────────────────────────────────────────
anthropic==0.42.0
openai==1.58.1

# ── Document Parsing (L2) ────────────────────────────
pypdf==5.1.0
python-docx==1.1.2

# ── Pipeline (L3) ────────────────────────────────────
langgraph==0.2.60
langchain-core==0.3.28

# ── Multi-Agent Debate (L6) ──────────────────────────
crewai>=0.86.0,<1.0.0

# ── HTTP Client ───────────────────────────────────────
httpx==0.28.1

# ── Testing ───────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.0
aiosqlite==0.20.0
